refactor(types): move adapters types back to adapters/ directory
