+++ packages/next-safe-action/src/hooks-utils.ts
@@ -1,7 +1,7 @@
 import * as React from "react";
 import {} from "react/experimental";
 import type {} from "zod";
-import type { InferIn, Schema } from "./adapters.types";
+import type { InferIn, Schema } from "./adapters/types";
 import type { HookActionStatus, HookCallbacks, HookResult } from "./hooks.types";
 
 export const getActionStatus = <
